requirements.txt
lxml
python-dateutil>=2.1
unittest2>=0.5.1
urllib3
jinja
